group tests and put into separate files
