+++ tests/tests.js
@@ -1,24 +1,8 @@
 'use strict';
 
-/* global describe it before */
-
-const chakram = require('chakram'),
-  expect = chakram.expect,
-  moment = require('moment');
-
 process.env.OSEM_TEST_BASE_URL = 'http://localhost:8000';
 
-const BASE_URL = process.env.OSEM_TEST_BASE_URL,
-  findAllSchema = require('./data/findAllSchema'),
-  csv_example_data = require('./data/csv_example_data'),
-  json_submit_data = require('./data/json_submit_data'),
-  byte_submit_data = require('./data/byte_submit_data'),
-  publishMqttMessage = require('./helpers/mqtt');
-
-const path = require('path').join(__dirname, 'routes');
-
-// must run at the start, b/c tests assume an empty database
-require('./location_tests');
+const path = require('path').join(__dirname, 'tests');
 
 require('fs')
   .readdirSync(path)
@@ -27,611 +11,3 @@ require('fs')
     require(`${path}/${file}`);
     /* eslint-enable global-require */
   });
-
-describe('openSenseMap API', function () {
-  let jwt;
-  let boxes = {};
-  let boxIds = [];
-
-  const boxCount = 3;
-
-  before(function () {
-    return chakram.post(`${BASE_URL}/users/sign-in`, { email: 'tester2@test.test', password: '12345678910' })
-      .then(function (response) {
-        expect(response).to.have.status(200);
-        expect(response).to.have.header('content-type', 'application/json; charset=utf-8');
-
-        expect(response.body.token).to.exist;
-
-        jwt = response.body.token;
-
-        return chakram.get(`${BASE_URL}/users/me/boxes`, { headers: { 'Authorization': `Bearer ${jwt}` } });
-      })
-      .then(function (response) {
-        boxIds = response.body.data.boxes.map(b => b._id);
-        boxes = response.body.data.boxes;
-
-        return chakram.wait();
-      });
-  });
-
-  describe('/boxes', function () {
-    let countMeasurements = 0;
-
-    it('should accept a single measurement via POST', function () {
-      let submitTime;
-
-      return chakram.post(`${BASE_URL}/boxes/${boxIds[0]}/${boxes[0].sensors[0]._id}`, { 'value': 312.1 })
-        .then(function (response) {
-          submitTime = moment.utc(response.response.headers.date, 'ddd, DD MMM YYYY HH:mm:ss GMT');
-          expect(response).to.have.status(201);
-          expect(response.body).to.equal('Measurement saved in box');
-
-          return chakram.get(`${BASE_URL}/boxes/${boxIds[0]}`);
-        })
-        .then(function (response) {
-          expect(response).to.have.json('sensors', function (sensors) {
-            sensors.forEach(function (sensor) {
-              if (sensor._id === boxes[0].sensors[0]._id) {
-                expect(sensor.lastMeasurement).not.to.be.null;
-                expect(sensor.lastMeasurement.createdAt).to.exist;
-                const createdAt = moment.utc(sensor.lastMeasurement.createdAt);
-                expect(createdAt.diff(submitTime)).to.be.below(1000);
-                expect(sensor.lastMeasurement.value).to.be.equal('312.1');
-              }
-            });
-          });
-          countMeasurements = countMeasurements + 1;
-
-          return chakram.wait();
-        });
-    });
-
-    it('should accept a single measurement with timestamp via POST', function () {
-      const submitTime = moment.utc().toISOString();
-
-      return chakram.post(`${BASE_URL}/boxes/${boxIds[0]}/${boxes[0].sensors[1]._id}`, { 'value': 123.4, 'createdAt': submitTime })
-        .then(function (response) {
-          expect(response).to.have.status(201);
-          expect(response.body).to.equal('Measurement saved in box');
-
-          return chakram.get(`${BASE_URL}/boxes/${boxIds[0]}`);
-        })
-        .then(function (response) {
-          expect(response).to.have.json('sensors', function (sensors) {
-            sensors.forEach(function (sensor) {
-              if (sensor._id === boxes[0].sensors[1]._id) {
-                expect(sensor.lastMeasurement).not.to.be.null;
-                expect(sensor.lastMeasurement.createdAt).to.exist;
-                const createdAt = moment.utc(sensor.lastMeasurement.createdAt);
-                expect(createdAt.diff(submitTime)).to.be.equal(0);
-                expect(sensor.lastMeasurement.value).to.be.equal('123.4');
-              }
-            });
-          });
-          countMeasurements = countMeasurements + 1;
-
-          return chakram.wait();
-        });
-    });
-
-    it('should reject a single measurement with timestamp too far into the future via POST', function () {
-      const submitTime = moment.utc().add(1.5, 'minutes')
-        .toISOString();
-
-      return chakram.post(`${BASE_URL}/boxes/${boxIds[0]}/${boxes[0].sensors[1]._id}`, { 'value': 123.4, 'createdAt': submitTime })
-        .then(function (response) {
-          expect(response).to.have.status(422);
-
-          return chakram.wait();
-        });
-    });
-
-    it('should accept multiple measurements as csv via POST', function () {
-      let submitTime;
-
-      return chakram.post(`${BASE_URL}/boxes/${boxIds[0]}/data`, csv_example_data.no_timestamps(boxes[0].sensors), { json: false, headers: { 'content-type': 'text/csv' } })
-        .then(function (response) {
-          submitTime = moment.utc(response.response.headers.date, 'ddd, DD MMM YYYY HH:mm:ss GMT');
-          expect(response).to.have.status(201);
-
-          return chakram.get(`${BASE_URL}/boxes/${boxIds[0]}`);
-        })
-        .then(function (response) {
-          expect(response).to.have.json('sensors', function (sensors) {
-            sensors.forEach(function (sensor) {
-              expect(sensor.lastMeasurement).not.to.be.null;
-              expect(sensor.lastMeasurement.createdAt).to.exist;
-              const createdAt = moment.utc(sensor.lastMeasurement.createdAt);
-              expect(createdAt.diff(submitTime)).to.be.below(1000);
-            });
-          });
-          countMeasurements = countMeasurements + boxes[0].sensors.length;
-
-          return chakram.wait();
-        });
-    });
-
-    it('should accept multiple measurements with timestamps as csv via POST', function () {
-      let submitTime;
-
-      return chakram.post(`${BASE_URL}/boxes/${boxIds[0]}/data`, csv_example_data.with_timestamps(boxes[0].sensors), { json: false, headers: { 'content-type': 'text/csv' } })
-        .then(function (response) {
-          submitTime = moment.utc(response.response.headers.date, 'ddd, DD MMM YYYY HH:mm:ss GMT');
-          expect(response).to.have.status(201);
-
-          return chakram.get(`${BASE_URL}/boxes/${boxIds[0]}`);
-        })
-        .then(function (response) {
-          expect(response).to.have.json('sensors', function (sensors) {
-            sensors.forEach(function (sensor) {
-              expect(sensor.lastMeasurement).not.to.be.null;
-              expect(sensor.lastMeasurement.createdAt).to.exist;
-              const createdAt = moment.utc(sensor.lastMeasurement.createdAt);
-              expect(submitTime.diff(createdAt, 'minutes')).to.be.below(5);
-            });
-          });
-          countMeasurements = countMeasurements + boxes[0].sensors.length;
-
-          return chakram.wait();
-        });
-    });
-
-    it('should reject multiple measurements with timestamps too far into the future as csv via POST', function () {
-      return chakram.post(`${BASE_URL}/boxes/${boxIds[0]}/data`, csv_example_data.with_timestamps_future(boxes[0].sensors), { json: false, headers: { 'content-type': 'text/csv' } })
-        .then(function (response) {
-          expect(response).to.have.status(422);
-
-          return chakram.wait();
-        });
-    });
-
-    it('should reject multiple measurements with too many fields as csv via POST', function () {
-      return chakram.post(`${BASE_URL}/boxes/${boxIds[0]}/data`, csv_example_data.with_too_many(boxes[0].sensors), { json: false, headers: { 'content-type': 'text/csv' } })
-        .then(function (response) {
-          expect(response).to.have.status(422);
-
-          return chakram.wait();
-        });
-    });
-
-
-    it('should accept multiple measurements with timestamps as json object via POST', function () {
-      let submitTime;
-
-      return chakram.post(`${BASE_URL}/boxes/${boxIds[0]}/data`, json_submit_data.json_obj(boxes[0].sensors))
-        .then(function (response) {
-          submitTime = moment.utc(response.response.headers.date, 'ddd, DD MMM YYYY HH:mm:ss GMT');
-          expect(response).to.have.status(201);
-          expect(response.body).to.equal('Measurements saved in box');
-
-          return chakram.get(`${BASE_URL}/boxes/${boxIds[0]}`);
-        })
-        .then(function (response) {
-          expect(response).to.have.json('sensors', function (sensors) {
-            sensors.forEach(function (sensor) {
-              expect(sensor.lastMeasurement).not.to.be.null;
-              expect(sensor.lastMeasurement.createdAt).to.exist;
-              const createdAt = moment.utc(sensor.lastMeasurement.createdAt);
-              expect(submitTime.diff(createdAt, 'minutes')).to.be.below(4);
-            });
-          });
-          countMeasurements = countMeasurements + boxes[0].sensors.length;
-
-          return chakram.wait();
-        });
-    });
-
-    it('should accept multiple measurements with timestamps as json array via POST', function () {
-      let submitTime;
-
-      return chakram.post(`${BASE_URL}/boxes/${boxIds[0]}/data`, json_submit_data.json_arr(boxes[0].sensors))
-        .then(function (response) {
-          submitTime = moment.utc(response.response.headers.date, 'ddd, DD MMM YYYY HH:mm:ss GMT');
-          expect(response).to.have.status(201);
-          expect(response.body).to.equal('Measurements saved in box');
-
-          return chakram.get(`${BASE_URL}/boxes/${boxIds[0]}`);
-        })
-        .then(function (response) {
-          expect(response).to.have.json('sensors', function (sensors) {
-            sensors.forEach(function (sensor) {
-              expect(sensor.lastMeasurement).not.to.be.null;
-              expect(sensor.lastMeasurement.createdAt).to.exist;
-              const createdAt = moment.utc(sensor.lastMeasurement.createdAt);
-              expect(submitTime.diff(createdAt, 'minutes')).to.be.below(4);
-            });
-          });
-          countMeasurements = countMeasurements + boxes[0].sensors.length;
-
-          return chakram.wait();
-        });
-    });
-
-    it('should accept multiple measurements as bytes via POST', function () {
-      let submitTime;
-
-      return chakram.post(`${BASE_URL}/boxes/${boxIds[0]}/data`, byte_submit_data(boxes[0].sensors), { json: false, headers: { 'content-type': 'application/sbx-bytes' } })
-        .then(function (response) {
-          submitTime = moment.utc(response.response.headers.date, 'ddd, DD MMM YYYY HH:mm:ss GMT');
-          expect(response).to.have.status(201);
-          expect(response.body).to.equal('"Measurements saved in box"');
-
-          return chakram.get(`${BASE_URL}/boxes/${boxIds[0]}`);
-        })
-        .then(function (response) {
-          expect(response).to.have.json('sensors', function (sensors) {
-            sensors.forEach(function (sensor) {
-              expect(sensor.lastMeasurement).not.to.be.null;
-              expect(sensor.lastMeasurement.createdAt).to.exist;
-              const createdAt = moment.utc(sensor.lastMeasurement.createdAt);
-              expect(submitTime.diff(createdAt, 'minutes')).to.be.below(4);
-            });
-          });
-          countMeasurements = countMeasurements + boxes[0].sensors.length;
-
-          return chakram.wait();
-        });
-    });
-
-    it('should accept multiple measurements as bytes with timestamp via POST', function () {
-      let submitTime;
-
-      return chakram.post(`${BASE_URL}/boxes/${boxIds[0]}/data`, byte_submit_data(boxes[0].sensors, true), { json: false, headers: { 'content-type': 'application/sbx-bytes-ts' } })
-        .then(function (response) {
-          submitTime = moment.utc(response.response.headers.date, 'ddd, DD MMM YYYY HH:mm:ss GMT');
-          expect(response).to.have.status(201);
-          expect(response.body).to.equal('"Measurements saved in box"');
-
-          return chakram.get(`${BASE_URL}/boxes/${boxIds[0]}`);
-        })
-        .then(function (response) {
-          expect(response).to.have.json('sensors', function (sensors) {
-            sensors.forEach(function (sensor) {
-              expect(sensor.lastMeasurement).not.to.be.null;
-              expect(sensor.lastMeasurement.createdAt).to.exist;
-              const createdAt = moment.utc(sensor.lastMeasurement.createdAt);
-              expect(submitTime.diff(createdAt, 'minutes')).to.be.below(5);
-            });
-          });
-          countMeasurements = countMeasurements + boxes[0].sensors.length;
-
-          return chakram.wait();
-        });
-    });
-
-    it('should return /stats correctly', function () {
-      return chakram.get(`${BASE_URL}/stats`, { headers: { 'x-apicache-bypass': true } })
-        .then(function (response) {
-          const [boxes, measurements] = response.body;
-          expect(boxes).to.equal(boxCount);
-          expect(measurements).to.equal(countMeasurements);
-
-          return chakram.wait();
-        });
-    });
-
-    it('should allow download data through /boxes/:boxid/data/:sensorid', function () {
-      return chakram.get(`${BASE_URL}/boxes/${boxIds[0]}/data/${boxes[0].sensors[0]._id}`)
-        .then(function (response) {
-          expect(response).to.have.status(200);
-          expect(Array.isArray(response.body)).to.be.true;
-          expect(response).to.have.header('content-type', 'application/json; charset=utf-8');
-          expect(response.body.length).to.be.above(4);
-
-          return chakram.wait();
-        });
-    });
-
-    it('should allow download data through /boxes/:boxid/data/:sensorid as csv', function () {
-      return chakram.get(`${BASE_URL}/boxes/${boxIds[0]}/data/${boxes[0].sensors[1]._id}?format=csv`)
-        .then(function (response) {
-          expect(response).to.have.status(200);
-          expect(response.body).not.to.be.empty;
-          expect(response).to.have.header('content-type', 'text/csv');
-          expect(response).to.have.header('Content-Disposition', `attachment; filename=${boxes[0].sensors[1]._id}.csv`);
-
-          return chakram.wait();
-        });
-    });
-
-    it('should allow download data through /boxes/data/:sensorid as csv', function () {
-      return chakram.get(`${BASE_URL}/boxes/data/?boxid=${boxIds[0]}&phenomenon=Temperatur`)
-        .then(function (response) {
-          expect(response).to.have.status(200);
-          expect(response.body).not.to.be.empty;
-          expect(response).to.have.header('content-type', 'text/csv');
-
-          return chakram.wait();
-        });
-    });
-
-    it('should allow download data via POST through /boxes/data/:sensorid as csv', function () {
-      return chakram.post(`${BASE_URL}/boxes/data`, { boxid: boxIds[0], phenomenon: 'Temperatur' })
-        .then(function (response) {
-          expect(response).to.have.status(200);
-          expect(response.body).not.to.be.empty;
-          expect(response).to.have.header('content-type', 'text/csv');
-
-          return chakram.wait();
-        });
-    });
-
-    it('should multiple csv measurements', function () {
-      return chakram.post(`${BASE_URL}/boxes/${boxIds[1]}/data`, csv_example_data.ten_days_ago_many(boxes[1].sensors), { json: false, headers: { 'content-type': 'text/csv' } })
-        .then(function (response) {
-          expect(response).to.have.status(201);
-
-          return chakram.get(`${BASE_URL}/boxes/${boxIds[1]}`);
-        })
-        .then(function (response) {
-          expect(response).to.have.json('sensors', function (sensors) {
-            sensors.forEach(function (sensor) {
-              expect(sensor.lastMeasurement).not.to.be.null;
-              expect(sensor.lastMeasurement.createdAt).to.exist;
-            });
-          });
-
-          return chakram.wait();
-        });
-    });
-
-    it('should return the correct count and correct schema of boxes for /boxes GET', function () {
-      return chakram.get(`${BASE_URL}/boxes`)
-        .then(function (response) {
-          expect(response).to.have.status(200);
-          expect(Array.isArray(response.body)).to.be.true;
-          expect(response.body.length).to.be.equal(boxCount);
-          expect(response).to.have.header('content-type', 'application/json; charset=utf-8');
-          expect(response).to.have.schema(findAllSchema);
-
-          return chakram.wait();
-        });
-    });
-
-    it('should return the correct count and correct schema of boxes for /boxes GET with date parameter', function () {
-      const ten_days_ago = moment.utc().subtract(10, 'days');
-
-      return chakram.get(`${BASE_URL}/boxes?date=${ten_days_ago.toISOString()}`)
-        .then(function (response) {
-          expect(response).to.have.status(200);
-          expect(Array.isArray(response.body)).to.be.true;
-          expect(response.body.length).to.be.equal(1);
-          expect(response).to.have.header('content-type', 'application/json; charset=utf-8');
-          expect(response).to.have.schema(findAllSchema);
-          expect(response.body[0].sensors.some(function (sensor) {
-            return moment.utc(sensor.lastMeasurement.createdAt).diff(ten_days_ago) < 10;
-          })).to.be.true;
-
-          return chakram.wait();
-        });
-    });
-
-    it('should return the correct count and correct schema of boxes for /boxes GET with date parameter outside of data range', function () {
-      const eleven_days_ago = moment.utc().subtract(11, 'days');
-
-      return chakram.get(`${BASE_URL}/boxes?date=${eleven_days_ago.toISOString()}`)
-        .then(function (response) {
-          expect(response).to.have.status(200);
-          expect(Array.isArray(response.body)).to.be.true;
-          expect(response.body.length).to.be.equal(0);
-          expect(response).to.have.header('content-type', 'application/json; charset=utf-8');
-          expect(response).to.have.schema(findAllSchema);
-
-          return chakram.wait();
-        });
-    });
-
-    it('should return the correct count and correct schema of boxes for /boxes GET with two date parameters', function () {
-      const now = moment.utc();
-
-      return chakram.get(`${BASE_URL}/boxes?date=${now.clone().subtract(1, 'minute')
-        .toISOString()},${now.toISOString()}`)
-        .then(function (response) {
-          expect(response).to.have.status(200);
-          expect(Array.isArray(response.body)).to.be.true;
-          expect(response.body.length).to.be.equal(1);
-          expect(response).to.have.header('content-type', 'application/json; charset=utf-8');
-          expect(response).to.have.schema(findAllSchema);
-
-          return chakram.get(`${BASE_URL}/boxes?date=${now.clone().subtract(10, 'days')
-            .subtract(10, 'minutes')
-            .toISOString()},${now.toISOString()}`);
-        })
-        .then(function (response) {
-          expect(response).to.have.status(200);
-          expect(Array.isArray(response.body)).to.be.true;
-          expect(response.body.length).to.be.equal(2);
-          expect(response).to.have.header('content-type', 'application/json; charset=utf-8');
-          expect(response).to.have.schema(findAllSchema);
-
-          return chakram.wait();
-        });
-    });
-
-    it('should return the correct count and correct schema of boxes for /boxes GET with date parameter after deleted sensor', function () {
-      const now = moment.utc();
-
-      return chakram.get(`${BASE_URL}/boxes?date=${now.toISOString()}&phenomenon=Temperatur`)
-        .then(function (response) {
-          expect(response).to.have.status(200);
-          expect(Array.isArray(response.body)).to.be.true;
-          expect(response.body.length).to.be.equal(1);
-          expect(response).to.have.header('content-type', 'application/json; charset=utf-8');
-          expect(response).to.have.schema(findAllSchema);
-
-          return chakram.wait();
-        });
-    });
-
-    it('should return the correct count and correct schema of boxes for /boxes GET with date parameter after deleted sensor #2', function () {
-      const ten_days_ago = moment.utc().subtract(10, 'days');
-
-      return chakram.get(`${BASE_URL}/boxes?date=${ten_days_ago.toISOString()}&phenomenon=Luftdruck`)
-        .then(function (response) {
-          expect(response).to.have.status(200);
-          expect(Array.isArray(response.body)).to.be.true;
-          expect(response.body.length).to.be.equal(1);
-          expect(response).to.have.header('content-type', 'application/json; charset=utf-8');
-          expect(response).to.have.schema(findAllSchema);
-
-          return chakram.wait();
-        });
-    });
-
-    it('should return the correct count and correct schema of boxes for /boxes GET with two date parameters after deleted sensor', function () {
-      const now = moment.utc();
-
-      return chakram.get(`${BASE_URL}/boxes?date=${now.clone().subtract(5, 'minutes')
-        .toISOString()},${now.toISOString()}&phenomenon=Temperatur`)
-        .then(function (response) {
-          expect(response).to.have.status(200);
-          expect(Array.isArray(response.body)).to.be.true;
-          expect(response.body.length).to.be.equal(1);
-          expect(response).to.have.header('content-type', 'application/json; charset=utf-8');
-          expect(response).to.have.schema(findAllSchema);
-
-          return chakram.wait();
-        });
-    });
-
-    it('should accept measurements through mqtt', function () {
-      const submitTime = moment.utc();
-
-      const payload = JSON.stringify(json_submit_data.json_arr(boxes[0].sensors));
-
-      return publishMqttMessage('mqtt://mosquitto', 'mytopic', payload)
-        .then(function () {
-          return new Promise(function (resolve) {
-            setTimeout(resolve, 500);
-          })
-            .then(function () {
-              return chakram.get(`${BASE_URL}/boxes/${boxIds[0]}`);
-            });
-        })
-        .then(function (response) {
-          expect(response).to.have.status(200);
-          expect(response).to.have.json('sensors', function (sensors) {
-            sensors.forEach(function (sensor) {
-              expect(sensor.lastMeasurement).not.to.be.null;
-              expect(sensor.lastMeasurement.createdAt).to.exist;
-              const createdAt = moment.utc(sensor.lastMeasurement.createdAt);
-              expect(submitTime.diff(createdAt, 'minutes')).to.be.below(5);
-            });
-          });
-          countMeasurements = countMeasurements + boxes[0].sensors.length;
-
-          return chakram.wait();
-        });
-    });
-
-    it('should allow to delete all data for a single sensor', function () {
-      const tempsensor_id = boxes[1].sensors[boxes[1].sensors.findIndex(s => s.title === 'Temperatur')]._id;
-      const payload = { 'deleteAllMeasurements': true };
-
-      return chakram.delete(`${BASE_URL}/boxes/${boxIds[1]}/${tempsensor_id}/measurements`, payload, { headers: { 'content-type': 'application/json', 'Authorization': `Bearer ${jwt}` } })
-        .then(function (response) {
-          expect(response).to.have.status(200);
-
-          return chakram.get(`${BASE_URL}/boxes/${boxIds[1]}/data/${tempsensor_id}`);
-        })
-        .then(function (response) {
-          expect(Array.isArray(response.body)).to.be.true;
-          expect(response.body.length).to.equal(0);
-
-          return chakram.get(`${BASE_URL}/boxes/${boxIds[1]}`);
-        })
-        .then(function (response) {
-          expect(response).to.have.json('sensors', function (sensors) {
-            sensors.forEach(function (sensor) {
-              if (sensor._id === tempsensor_id) {
-                expect(sensor.lastMeasurement).not.to.exist;
-              }
-            });
-          });
-
-          return chakram.wait();
-        });
-    });
-
-    it('should allow to delete data for a single sensor through specifying timestamps', function () {
-      let sensor_id;
-      const payload = { timestamps: [] };
-
-      return chakram.get(`${BASE_URL}/boxes/${boxIds[1]}/sensors`)
-        .then(function (response) {
-          expect(response).to.have.status(200);
-          expect(response.body.sensors).to.not.be.empty;
-          expect(Array.isArray(response.body.sensors)).to.be.true;
-          const sensor = response.body.sensors.find(s => s.title === 'Windgeschwindigkeit');
-          expect(sensor).to.not.be.undefined;
-          expect(sensor.lastMeasurement).to.not.be.undefined;
-          sensor_id = sensor._id;
-          payload.timestamps.push(sensor.lastMeasurement.createdAt);
-
-          return chakram.delete(`${BASE_URL}/boxes/${boxIds[1]}/${sensor_id}/measurements`, payload, { headers: { 'content-type': 'application/json', 'Authorization': `Bearer ${jwt}` } });
-        })
-        .then(function (response) {
-          expect(response).to.have.status(200);
-
-          return chakram.get(`${BASE_URL}/boxes/${boxIds[1]}/data/${sensor_id}`);
-        })
-        .then(function (response) {
-          expect(Array.isArray(response.body)).to.be.true;
-          expect(response.body).to.be.empty;
-
-          return chakram.get(`${BASE_URL}/boxes/${boxIds[1]}`);
-        })
-        .then(function (response) {
-          expect(response).to.have.json('sensors', function (sensors) {
-            sensors.forEach(function (sensor) {
-              if (sensor._id === sensor_id) {
-                expect(sensor.lastMeasurement.createdAt).to.not.equal(payload.timestamps[0]);
-              }
-            });
-          });
-
-          return chakram.wait();
-        });
-    });
-
-    it('should allow to delete all data for a single sensor through specifying from-date and to-date', function () {
-      const sensor_id = boxes[1].sensors[boxes[1].sensors.findIndex(s => s.title === 'Luftdruck')]._id;
-      const payload = { 'from-date': moment.utc().subtract(1, 'year')
-        .toISOString(), 'to-date': moment.utc().toISOString() };
-
-      return chakram.delete(`${BASE_URL}/boxes/${boxIds[1]}/${sensor_id}/measurements`, payload, { headers: { 'content-type': 'application/json', 'Authorization': `Bearer ${jwt}` } })
-        .then(function (response) {
-          expect(response).to.have.status(200);
-
-          return chakram.get(`${BASE_URL}/boxes/${boxIds[1]}/data/${sensor_id}`);
-        })
-        .then(function (response) {
-          expect(Array.isArray(response.body)).to.be.true;
-          expect(response.body.length).to.equal(0);
-
-          return chakram.get(`${BASE_URL}/boxes/${boxIds[1]}`);
-        })
-        .then(function (response) {
-          expect(response).to.have.json('sensors', function (sensors) {
-            sensors.forEach(function (sensor) {
-              if (sensor._id === sensor_id) {
-                expect(sensor.lastMeasurement).not.to.exist;
-              }
-            });
-          });
-
-          return chakram.wait();
-        });
-    });
-  });
-
-});
-
-require('./feinstaubaddon_tests');
-
-require('./luftdaten_tests');
-
-require('./delete_user_tests');
-
-require('./mail_tests');
