+++ tests/tests/010-luftdaten-test.js
@@ -7,8 +7,8 @@ const chakram = require('chakram'),
   moment = require('moment');
 
 const BASE_URL = process.env.OSEM_TEST_BASE_URL,
-  valid_sensebox = require('./data/valid_sensebox'),
-  luftdaten_example_data = require('./data/luftdaten_example_data');
+  valid_sensebox = require('../data/valid_sensebox'),
+  luftdaten_example_data = require('../data/luftdaten_example_data');
 
 describe('openSenseMap API luftdaten.info devices', function () {
   let jwt, dht11_id, dht22_id, bmp180_id, bme280_id, custom_id;
